+++ lib/src/web_tokens/base/color/scrim/scrim_base.dart
@@ -1,0 +1,14 @@
+/*
+ * SPDX-FileCopyrightText: Copyright 2025 LG Electronics Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import 'package:flutter/material.dart';
+
+abstract class ScrimBase {
+  const ScrimBase();
+
+  Color get defaultColor;
+  Color get white;
+  Color get black;
+}